LEAF-78 Fixed pollen card, made it editable and refactored the code so its more readable
